refactor(list): use size_t

+++ src/linked_list.h
@@ -1,5 +1,7 @@
 #pragma once
 #include <stdbool.h>
+// Needed for size_t.
+#include <stdlib.h>
 
 /**
  * @file linked_list.h
@@ -65,7 +67,7 @@ bool ioopm_linked_list_contains(ioopm_list_t *list, int element);
 /// @brief Lookup the number of elements in the linked list in O(1) time
 /// @param list the linked list
 /// @return the number of elements in the list
-int ioopm_linked_list_size(ioopm_list_t *list);
+size_t ioopm_linked_list_size(ioopm_list_t *list);
 
 /// @brief Test whether a list is empty or not
 /// @param list the linked list
